ptp: Fix PTP_DFL_TOD dependency on FPGA_DFL
The Intel FPGA DFL ToD device driver is a driver that exposes the Intel
FPGA ToD IP as a PTP Hardware Clock device to the Linux PTP stack. This
driver implicitly takes a dependency on the FPGA Device Feature List
(DFL) Framework, as it uses the `module_dfl_driver` function, which
requires the `dfl_driver_register` and `dfl_driver_unregister`
symbols. This change corrects that by adding an explicit dependency on
the `FPGA_DFL` dependency in the `PTP_DFL_TOD` config.