[fb-only] build: Add selftestsconfig build target
Summary:
The kernel build system currently exposes a `kernel` BUCK macro that can
be used to generate targets to build a kernel version with some
specified arch, flavor, and debugging. We want to enable the ability to
run selftests for any of these flavors, but to do so, we need to include
config options specified by those selftests and build the kernel with
those config options.

When a kselftest requires config options, the contract is that it should
specify all of the options in a `config` file in the test directory.
Livepatch, for example, has the following configs:

```
CONFIG_LIVEPATCH=y
CONFIG_DYNAMIC_DEBUG=y
CONFIG_TEST_LIVEPATCH=m
```

And the bpf kselftests have many as well.

We'll eventually need to add build targets that build kernel seftests.
In order to do that, we'll also need to add a buildtarget that can
append the selftests `.config` options to the `.config` file created for
the entire build. This change therefore adds a new
`//facebook/scripts/selftestconfig` BUCK target that aggregates all of
the required selftest config options, and outputs them. A follow-on diff
will leverage this target to create a full `-selftests` kernel target.

Signed-off-by: David Vernet <void@manifault.com>

Test Plan:
`[dvernet@dev0025.ash9 ~/local/linux] NO_BUCKD=1 ./facebook/build/buck
run //facebook/scripts:selftestsconfig`: P476935182

And if we add another selftest to the list, such as `bpf`: P476935626

Also, in a follow-on change, verify that we can now run livepatch
selftest on a -selftests build: P478119125

Reviewers: #kernel_ops, #kernel-validation-and-deployment

Subscribers:

Tasks:

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D33846044