[fb-only] build: Package selftests in rpms
Summary:
In a set of prior changes (D33846056 and D33846044), we updated the
kernel build system to create -selftests BUCK targets that enable users
to compile kernels with the necessary build configs to run kselftests.
In order to actually invoke those tests, we'll need to provide users
with the ability to actually invoke the test suite at runtime. To do so,
we'll need to install the test, and export the bundled test suite as
part of the RPMs we ship.

This change therefore updates our fb-kernel.spec.template file to
install the test-suite, and bundle an executable from which a user can
invoke the selftests.

Signed-off-by: David Vernet <void@manifault.com>

Test Plan:
Verify that we can now run livepatch selftest on a -selftests build: P478119125
Also tested building and booting a normal build: P478127697

Reviewers: #kernel_ops, #kernel-validation-and-deployment

Subscribers:

Tasks:

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D33980523