[fb-only] Configure selftests in a single build variable
Summary:
In a set of prior changes (D33980523, D33846056, D33846044), we added
support for building -selftests targets which compile the kernel with
the config options required to run Linux kernel selftests. There are
currently two places where we specify the selftest suites that we want
to install and bundle -- one is where we invoke the selftestsconfig
script, and the other when we bundle the selftests rpm that will contain
the actual tests that are invoked by the end user.

So as to avoid us having the suites defined in two places, this change
updates the build process to instead define a global variable that is
passed as a parameter to both of the aforementioned consumers.

Signed-off-by: David Vernet <void@manifault.com>

Test Plan:
Verify the `selftestsconfig` script works as expected:
`[dvernet@dev0025.ash9 ~/local/linux] NO_BUCKD=1 ./facebook/build/buck
  run //facebook/scripts:selftestsconfig -- livepatch bpf`: P478144267

Verify that building selftests still works as expected:

Reviewers: #kernel_ops, #kernel-validation-and-deployment

Subscribers: 

Tasks: 

Tags: 


Differential Revision: https://phabricator.intern.facebook.com/D33985061