[fb-only] Generate -tar-pkg target that creates rpm + uname tarball
Summary:
In the `kernel` buck build macro, the final / top-level target creates a
set of RPMs that can be zipped into a `blobby.tgz` tarball, and installed
on a host with `kernelctl`. Tarring these RPMs seems like a common
operation. So as to save users the trouble of having to manually do it
themselves, this diff adds a new target that suffixes the existing
top-level target with `-tar-pkg`, and automatically puts all of the RPMs
and the uname file into a tar-pkg that can be consumed by `kernelctl`.

Signed-off-by: David Vernet <void@manifault.com>

Test Plan:
Builds: P478989192, P478994031

Also created both -selftests and regular variants of the -tar-pkg target,
and installed them on a kernelres host using `kernelctl install`.

Reviewers: #kernel_ops, #kernel-validation-and-deployment

Subscribers:

Tasks:

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D34076842