nvme: don't allow IO generating allocations under queue freeze
This will potentially hang the system, if the allocation ends up
recursing back into the same device and then waiting for that IO
to complete.

There's still a slight risk of a pagefault being triggered on the
memory that the application passes in for a read based passthrough
command, but that is deemed small enough not to worry about for now.

Note that upstream and 5.12-fb do not suffer from this issue, as the
passthrough support was cleaned up to properly only freeze the queue
for the duration of the IO. 5.6-fb and friends do a lot more under
the passthrough start/end section, which is what is causing the
issue.

Signed-off-by: Jens Axboe <axboe@kernel.dk>