[FB ONLY] ipc namespaces: don't use the async workqueue for freeing
This is FB only because we want to actually fix the async workqueue
upstream.  e1eb26fa62d04ec0955432be1aa8722a97cb52e7 upstream punted
freeing of ipc namespace objects to an async workqueue.  This was done
because we need to free some vfs mount related structures, which
involves a call to synchronize_rcu().  In hopes of avoiding that
latency, they switched to a workqueue instead.

But, there's no throttling of new allocations based on the size of the
workqueue list, and since synchronize_rcu() is slower than new namespace
creation, it's possible to end up with hundreds of thousands of frees
queued up.  Eventually we blow through the namespace limit and creation
returns ENOSPC.

Thread for the real fixes upstream:
https://lore.kernel.org/lkml/8E281DF1-248F-4861-A3C0-2573A5EFEE61@fb.com/

This commit is basically a revert of e1eb26fa62d04e, but I kept the
workqueue structures to make it live patchable.

Signed-off-by: Chris Mason <clm@fb.com>