[fb-only] Add diff and continuous selftests builds
Summary:
In a series of recent changes, we added support for building -selftests
variants of kernels. Additionally, in a recent www change, we enabled a
'selftests' build mode for the SandcastleKernelBuildCommand, which builds
the kernel in this -selftests mode, and also invokes a series of kernel
tests.

So as to get CI coverage of the -selftests build, this change adds UTD
directives to also build -selftests on diffs, and on a continuous basis.

Test Plan:
It doesn't seem possible to test anything with UTD. According to
https://www.internalfb.com/intern/utd/?run_id=cc9cf35e-d0e8-4ac3-8cb8-86f6a058cce8,
it looks like the -selftests rules were successfully parsed, and correctly
not scheduled due to it not being continuous. Conversely, the diff rule was
parsed and the job was successfully scheduled.

Reviewers: #kernel_ops, #kernel-validation-and-deployment

Subscribers:

Tasks:

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D34254106