[fb-only] Declare empty SKIP_TARGETS in selftests install
Summary:
We recently landed a set of diffs that allow us to build and run linux
kernel selftests on CI. Selftests are installed using the following
directive:

make -C tools/testing/selftests install TARGETS="..."

where one may list the set of selftest suites that they wish to have
installed in TARGETS. Confusingly, there is also an undocumented make
variable called SKIP_TARGETS that includes a list of test suites that
should be skipped by default. This variable exists because if TARGETS is
not specified, a large set of default targets are used, and SKIP_TARGETS
is basically the way to specify which targets (such as bpf) are
non-defaults.

While it's slightly non-trivial to add bpf to the suite of selftests due
to the fact that it requires some libraries that are not present in the
build container we use when making RPMs, we should still nullify
SKIP_TARGETS so that if someone wants to add bpf in the future, they're
not confused why it isn't working if they add it to the SELFTESTS
variable in facebook/config/defs.bzl.

Test Plan:
Verify that if we add 'bpf' to the SELFTESTS variable, that bpf is
included in the list of test suites that are attempted to be installed
as part of the RPM selftests subpackage: P480636312

Reviewers: #kernel_ops, #kernel-validation-and-deployment

Subscribers: 

Tasks: 

Tags: 


Differential Revision: https://phabricator.intern.facebook.com/D34272729