Add FAT messages to printk index
Summary:
In order for end users to quickly react to new issues that come up in
production, it is proving useful to leverage the printk indexing system. This
printk index enables kernel developers to use calls to printk() with changable
ad-hoc format strings, while still enabling end users to detect changes and
develop a semi-stable interface for detecting and parsing these messages.

So that detailed FAT messages are captured by this printk index, this patch
wraps fat_msg with a macro.

Test Plan:
Build kernels and use printk_cli to parse the index out and compare.
After adding this, 40 additional entries are present in the index: P483442957

Reviewers: alastorze, cdown

Subscribers:

Tasks:

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D34533587